feat(functional-split): ver1.0
